==> r_scripts/Markdown_scripts/5.PCA_Heatmap_field_lab_GENE.rmd <==
---
title: "PCA_Heatmap_field_lab_GENE"
author: "Fay"
date: '2022-11-08'
output:
  pdf_document:
    keep_md: yes 
    fig_width: 12
    fig_height: 8
  html_document:
    df_print: paged
---

# Load libraries

```{r}
library(tidyverse)
library(dplyr)
library(stringr)
library(FactoMineR)
library(reshape2)
library(corrplot)
library(factoextra)
library(lmtest)
library(ggpubr)
library(janitor)
library(pheatmap)
library(visdat)
```


# Load data
```{r}
hm <- read.csv("output_data/2.imputed_MICE_data_set.csv")
```

# vectors for selecting

```{r}
Gene_field   <- c("IFNy", "CXCR3", "IL.6", "IL.13",# "IL.10",
                "IL1RN","CASP1", "CXCL9", "IDO1", "IRGM1", "MPO", 
                "MUC2", "MUC5AC", "MYD88", "NCR1", "PRF1", "RETNLB", "SOCS1", 
                "TICAM1", "TNF") #"IL.12", "IRG6")

#add a suffix to represent changes in data file
Gene_field_imp <- paste(Gene_field, "imp", sep = "_")

Gene_lab   <- c("IFNy", "CXCR3", "IL.6", "IL.13",# "IL.10",
                "IL1RN","CASP1", "CXCL9", "IDO1", "IRGM1", "MPO", 
                "MUC2", "MUC5AC", "MYD88", "NCR1", "PRF1", "RETNLB", "SOCS1", 
                "TICAM1", "TNF") # "IL.12", "IRG6")

#add a suffix to represent changes in data file
Gene_lab_imp <- paste(Gene_lab, "imp", sep = "_")

Genes_wild   <- c("IFNy", "CXCR3", "IL.6", "IL.13", "IL.10", 
                  "IL1RN","CASP1", "CXCL9", "IDO1", "IRGM1", "MPO", 
                  "MUC2", "MUC5AC", "MYD88", "NCR1", "PRF1", "RETNLB", "SOCS1", 
                  "TICAM1", "TNF") #, "IL.12", "IRG6")

Genes_wild_imp <- paste(Genes_wild, "imp", sep = "_")

Facs_field <- c("Position", "CD4", "Treg", "Div_Treg", "Treg17", "Th1", 
                    "Div_Th1", "Th17", "Div_Th17", "CD8", "Act_CD8", 
                    "Div_Act_CD8", "IFNy_CD4", "IFNy_CD8","Treg_prop", 
                    "IL17A_CD4")  

Facs_field_imp <- paste(Facs_field, "imp", sep = "_")

Facs_wild <- c( "Treg", "CD4", "Treg17", "Th1", "Th17", "CD8",
                     "Act_CD8", "IFNy_CD4", "IL17A_CD4", "IFNy_CD8")

Facs_wild_imp <- paste(Facs_wild, "imp", sep = "_")
```

# PCA on the field genes -*imputed*
```{r}
###########field


field <- hm %>%
  dplyr::filter(origin == "Field") 

field <- unique(field)

#make a factor out of the melting curves (important for later visualization)
field <- field %>%
  dplyr::mutate(MC.Eimeria = as.factor(MC.Eimeria))

genes_mouse <- field %>%
  dplyr::select(c(Mouse_ID, "IFNy", "CXCR3", "IL.6", "IL.13", #"IL.10", 
                  "IL1RN","CASP1", "CXCL9", "IDO1", "IRGM1", "MPO", 
                  "MUC2", "MUC5AC", "MYD88", "NCR1", "PRF1", "RETNLB", "SOCS1", 
                  "TICAM1", "TNF")) 


genes <- genes_mouse %>%
  dplyr::select(-Mouse_ID)


#remove rows with only nas
genes <- genes[,colSums(is.na(genes))<nrow(genes)]

#remove colums with only nas 
genes <- genes[rowSums(is.na(genes)) != ncol(genes), ]


##select same rows in the first table
field <- field[row.names(genes), ]


###############lab
#select the genes and lab muce
lab <- hm %>%
  dplyr::filter(origin == "Lab", Position == "mLN") #selecting for mln to avoid
# duplicates

lab <- unique(lab)

gene_lab <- lab %>%
  dplyr::select(c(Mouse_ID, "IFNy", "CXCR3", "IL.6", "IL.13", #"IL.10", 
                  "IL1RN","CASP1", "CXCL9", "IDO1", "IRGM1", "MPO", 
                  "MUC2", "MUC5AC", "MYD88", "NCR1", "PRF1", "RETNLB", "SOCS1", 
                  "TICAM1", "TNF")) %>%
  rename_with(~str_remove(., '_imp'))

genes_lab <- unique(gene_lab)

genes_lab <- genes_lab[, -1]

#remove rows with only nas
genes_lab <- genes_lab[,colSums(is.na(genes_lab))<nrow(genes_lab)]

#remove colums with only nas 
genes_lab <- genes_lab[rowSums(is.na(genes_lab)) != ncol(genes_lab), ]


#select same rows in the first table
gene_lab <- gene_lab[row.names(genes_lab), ]

##select same rows in the first table
lab <- lab[row.names(gene_lab), ]

hm_genes <- rbind(genes_lab, genes)

hm_selection_g <- rbind(lab, field)

# we can now run a normal pca on the complete data set
res.pca <- PCA(hm_genes)
```



```{r dimensions, include = FALSE, echo = FALSE, warnings = FALSE}
#Description of the dimensions
# We get a correlation between each variable and the first dimension
dimdesc(res.pca)
```

Caution: When imputing data, the percentages of inertia associated with the first dimensions will be overestimated.

Another problem: the imputed data are, when the pca is performed considered like real observations.
But they are estimations!! 

Visualizing uncertainty due to issing data:

--> mulrimple imputation: generate several plausible values for each missing data point

We here visualize the variability, that is uncertainty on the plane defined by two pca axes. 

```{r, echo = FALSE, include = FALSE}
str(res.pca)
```



```{r, echo = FALSE}


hm_selection_g$pc1 <- res.pca$ind$coord[, 1] # indexing the first column

hm_selection_g$pc2 <- res.pca$ind$coord[, 2]  # indexing the second column




#We also need to extract the data for the variable contributions to each of the pc axes.
pca.vars <- res.pca$var$coord %>% data.frame


pca.vars$vars <- rownames(pca.vars)

pca.vars.m <- melt(pca.vars, id.vars = "vars")

source("r_scripts/functions/circle_fun.R")

circ <- circleFun(c(0,0),2,npoints = 500)

```





```{r correlations_genes_dimensions, echo = FALSE}
#It’s possible to use the function corrplot() [corrplot package] to highlight 
#the most contributing variables for each dimension:
var.contrib <- res.pca$var$contrib
corrplot(var.contrib, is.corr=FALSE) 
```

The function fviz_contrib() [factoextra package] can be used to draw a bar plot of variable contributions. If your data contains many variables, you can decide to show only the top contributing variables. The R code below shows the top 10 variables contributing to the principal components:


```{r contr_var_pc_genes, echo = FALSE}
# Contributions of variables to PC1
fviz_contrib(res.pca, choice = "var", axes = 1, top = 18)

```


```{r}
# Contributions of variables to PC2
fviz_contrib(res.pca, choice = "var", axes = 2, top = 18)
```


The genes MPO, IL1RN, Interleukin 13, and MUC5AC are associated with various biological functions and have been studied in the context of different diseases and processes. Here's a brief overview of their known roles:

MPO (Myeloperoxidase):

MPO is an enzyme primarily expressed in neutrophils and plays a role in the innate immune response.
It is involved in the generation of reactive oxygen species (ROS) and the oxidation of various compounds, contributing to host defense against pathogens.
MPO has been implicated in inflammation, oxidative stress, and various diseases, including cardiovascular diseases, chronic obstructive pulmonary disease (COPD), and cancer.
IL1RN (Interleukin 1 Receptor Antagonist):

IL1RN codes for the interleukin 1 receptor antagonist, a protein that acts as a competitive inhibitor of interleukin 1 (IL-1) signaling.
IL-1 is a pro-inflammatory cytokine involved in immune responses and inflammation.
IL1RN helps regulate the inflammatory response by binding to IL-1 receptors and preventing IL-1 from exerting its pro-inflammatory effects.
Dysregulation of IL1RN has been associated with various inflammatory diseases, such as rheumatoid arthritis, inflammatory bowel disease, and cardiovascular disorders.
Interleukin 13 (IL-13):

IL-13 is a cytokine involved in immune responses, particularly in allergy and asthma.
It is primarily produced by T helper type 2 (Th2) cells and regulates several aspects of the immune system, including inflammation, cell differentiation, and immunoglobulin production.
IL-13 is implicated in allergic diseases, such as asthma, where it promotes airway hyperresponsiveness, mucus production, and tissue remodeling.
MUC5AC (Mucin 5AC):

MUC5AC encodes a gel-forming mucin protein primarily expressed in the respiratory tract.
Mucins are large glycoproteins that contribute to the protection and lubrication of mucosal surfaces.
MUC5AC is a major component of the mucus layer in the airways and plays a crucial role in airway defense mechanisms, trapping and clearing inhaled particles and pathogens.
Dysregulation of MUC5AC expression and mucin production is associated with respiratory diseases, including asthma, chronic bronchitis, and cystic fibrosis.
It's important to note that these genes may have additional roles and functions beyond the ones mentioned here, as ongoing research continues to uncover their complexities in various biological processes and diseases.





User
What is the factor that associates these genes?
ChatGPT
The factor that associates these genes is their involvement in inflammatory processes. Each of the genes MPO, IL1RN, Interleukin 13 (IL-13), and MUC5AC has been linked to inflammation in different contexts.

MPO is associated with inflammation through its role in the generation of reactive oxygen species (ROS) and its contribution to oxidative stress and tissue damage in inflammatory conditions. It is often used as a marker of inflammation in diseases such as cardiovascular diseases and chronic obstructive pulmonary disease (COPD).

IL1RN, also known as the interleukin 1 receptor antagonist, is a key regulator of interleukin 1 (IL-1) signaling. IL-1 is a pro-inflammatory cytokine involved in the initiation and amplification of the inflammatory response. IL1RN acts as a competitive inhibitor of IL-1, dampening the inflammatory cascade and preventing excessive inflammation.

IL-13 is a cytokine that plays a crucial role in the immune response, particularly in allergic and inflammatory diseases. It is associated with Th2-driven immune responses, promoting inflammation, mucus production, and tissue remodeling in diseases like asthma and allergic rhinitis.

MUC5AC, a mucin protein, is involved in airway defense and plays a role in the inflammatory response of the respiratory tract. Abnormal regulation of MUC5AC expression and mucin production can lead to excessive mucus production, airway obstruction, and chronic inflammation in respiratory conditions such as asthma and chronic bronchitis.

While these genes are individually associated with inflammation, it is worth noting that they can also interact and influence each other's expression and functions in complex ways within the context of specific diseases or biological processes.





User
Is MUC5AC also involved in the intestinal tract?
ChatGPT
Yes, MUC5AC is also involved in the intestinal tract. Mucins are a family of high-molecular-weight glycoproteins that are major components of the mucus layer that covers and protects the epithelial surfaces of various tissues, including the respiratory tract, gastrointestinal tract, and reproductive tract.

MUC5AC is one of the mucin genes expressed in the gastrointestinal tract, particularly in the stomach and the surface epithelium of the colon. In the intestine, MUC5AC expression is relatively low compared to other mucin genes such as MUC2, which is the predominant mucin in the intestine. However, studies have shown that MUC5AC expression can be upregulated in certain conditions such as inflammatory bowel disease (IBD), including Crohn's disease and ulcerative colitis.

In the context of the intestinal tract, the expression of MUC5AC may be associated with mucosal inflammation and the protective response to pathogens or other insults. The overexpression or aberrant regulation of MUC5AC in the intestine can impact mucus production, mucosal barrier function, and the interaction between the host and the gut microbiota. This, in turn, can influence the pathogenesis and progression of intestinal diseases.


```{r contr_var_pc1_2_genes, echo = FALSE}
fviz_contrib(res.pca, choice = "var", axes = 1:2, top = 18)
```
The red dashed line on the graph above indicates the expected average contribution. If the contribution of the variables were uniform, the expected value would be 1/length(variables) = 1/10 = 10%. For a given component, a variable with a contribution larger than this cutoff could be considered as important in contributing to the component.

Note that, the total contribution of a given variable, on explaining the variations retained by two principal components, say PC1 and PC2, is calculated as contrib = [(C1 * Eig1) + (C2 * Eig2)]/(Eig1 + Eig2), where

C1 and C2 are the contributions of the variable on PC1 and PC2, respectively
Eig1 and Eig2 are the eigenvalues of PC1 and PC2, respectively. Recall that eigenvalues measure the amount of variation retained by each PC.
In this case, the expected average contribution (cutoff) is calculated as follow: As mentioned above, if the contributions of the 10 variables were uniform, the expected average contribution on a given PC would be 1/10 = 10%. The expected average contribution of a variable for PC1 and PC2 is : [(10* Eig1) + (10 * Eig2)]/(Eig1 + Eig2)



```{r pca_contribution_genes, echo = FALSE}

#The most important (or, contributing) variables can be highlighted on the 
#correlation plot as follow:
fviz_pca_var(res.pca, col.var = "contrib",
             gradient.cols = c("#00AFBB", "#E7B800", "#FC4E07")
             )
```



To visualize the contribution of individuals to the first two principal components:

```{r contr_individuals_genes, echo = FALSE}
# Total contribution on PC1 and PC2
fviz_contrib(res.pca, choice = "ind", axes = 1:2)
```

PCA + Biplot combination

```{r pca_biplot_genes, echo = FALSE}


fviz_pca_biplot(res.pca, 
                col.ind = hm_selection_g$origin, palette = "jco", 
                addEllipses = TRUE, fieldel = "var",
                col.var = "black", repel = TRUE,
                legend.title = "Origin of mice") 

```
 In the following example, we want to color both individuals and variables by groups. The trick is to use pointshape = 21 for individual points. This particular point shape can be filled by a color using the argument fill.ind. The border line color of individual points is set to “black” using col.ind. To color variable by groups, the argument col.var will be used.





### repeating the heatmap on the now imputed data
```{r}
mouse_id <- hm_selection_g %>%
  dplyr::select(Mouse_ID)

hm_genes <- cbind(mouse_id, hm_genes)


# turn the data frame into a matrix and transpose it. We want to have each cell 
 # type as a row name 
 hm_genes <- t(as.matrix(hm_genes))
 
 # turn the first row into column names
 hm_genes %>%
     row_to_names(row_number = 1) -> heatmap_data
 
 heatmap_data <- as.data.frame(heatmap_data)
 
 table(rowSums(is.na(heatmap_data)) == nrow(heatmap_data))

 
# turn the columns to numeric other wise the heatmap function will not work
 heatmap_data[] <- lapply(heatmap_data, function(x) as.numeric(as.character(x)))

 # remove columns with only NAs 
 heatmap_data <- Filter(function(x)!all(is.na(x)), heatmap_data) 
 
 #remove rows with only Nas
 heatmap_data <-  heatmap_data[, colSums(is.na(heatmap_data)) != 
                                   nrow(heatmap_data)]
 
  

#Prepare the annotation data frame
annotation_df <- as_tibble(hm_selection_g) %>%
    dplyr::select(c("Mouse_ID", "origin")) 
  
annotation_df <- unique(annotation_df) 

annotation_df <- as.data.frame(annotation_df)

 ### Prepare the annotation columns for the heatmap

rownames(annotation_df) <- annotation_df$Mouse_ID

# Match the row names to the heatmap data frame
rownames(annotation_df) <- colnames(heatmap_data)

#remove the unecessary column
annotation_df <- annotation_df %>% dplyr::select(-Mouse_ID, )


```



Heatmap on gene expression data: 


```{r, echo = FALSE}
pheatmap(heatmap_data, annotation_col = annotation_df, scale = "row")
```
